LPS-99657 Class that extends FindStrutsAction should be in package *.struts
